refactor(documents): extract shared OpenAI transport/image helpers + resolve_api_key
